src/requirements.txt
Flask==0.12.2
requests==2.19.1
psycopg2-binary
